util: Move copy3 in utils with types, to avoid code duplication
